refactor: move get_iso_code to wizard/utils.py
